requirements.txt
selenium
kafka-python
pymongo
python-dotenv